requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
python-dotenv==1.0.0
requests==2.31.0
twilio==8.11.0
schedule==1.2.0